Knowledge refactor
